Do not require err_str for socket.error
When catching socket.error exception it is not guaranteed to have two
values, especially when the buffer is full. This commit adjust it and
does not require two variables to unpack.

Without this commit the following exception is raised:

Traceback (most recent call last):
  File "/usr/lib/python2.7/threading.py", line 810, in __bootstrap_inner
    self.run()
  File "/usr/lib/python2.7/threading.py", line 763, in run
    self.__target(*self.__args, **self.__kwargs)
  File "./le.py", line 1591, in run
    self._send_entry(entry)
  File "./le.py", line 1559, in _send_entry
    except socket.error, (err_no, err_str):
ValueError: need more than 1 value to unpack